Fix 9 digits number value, add validation when create/update voucher

+++ src/main/java/campaign/service/dto/RewardConditionRewardDTO.java
@@ -16,9 +16,9 @@ public class RewardConditionRewardDTO {
 
     private Integer prizeType;
 
-    private Float prizeValue;
+    private Long prizeValue;
 
-    private Float numOfPrize;
+    private Long numOfPrize;
 
     private Integer released;
 
@@ -102,19 +102,19 @@ public class RewardConditionRewardDTO {
         this.prizeType = prizeType;
     }
 
-    public Float getPrizeValue() {
+    public Long getPrizeValue() {
         return prizeValue;
     }
 
-    public void setPrizeValue(Float prizeValue) {
+    public void setPrizeValue(Long prizeValue) {
         this.prizeValue = prizeValue;
     }
 
-    public Float getNumOfPrize() {
+    public Long getNumOfPrize() {
         return numOfPrize;
     }
 
-    public void setNumOfPrize(Float numOfPrize) {
+    public void setNumOfPrize(Long numOfPrize) {
         this.numOfPrize = numOfPrize;
     }
